… Orm/Xtensive.Orm.Tests/Linq/InTest.cs
… Orm/Xtensive.Orm.Tests/Storage/TemporaryTablePopulationTest.cs
… Orm/Xtensive.Orm/Orm/IncludeAlgorithm.cs
… Orm/Xtensive.Orm/Orm/Providers/CommandProcessing/CommandFactory.cs
… Orm/Xtensive.Orm/Orm/Providers/ProviderFeatures.cs
… Orm/Xtensive.Orm/Orm/Providers/ProviderInfoBuilder.cs
… Orm/Xtensive.Orm/Orm/Providers/Requests/QueryRowFilterParameterBinding.cs
… Orm/Xtensive.Orm/Orm/Providers/SqlCompiler.Include.cs
… Orm/Xtensive.Orm/Orm/Providers/SqlCompiler.cs
… Orm/Xtensive.Orm/Orm/Providers/StorageDriver.cs
… Orm/Xtensive.Orm/Orm/Upgrade/Internals/SqlAsyncWorker.cs
… Orm/Xtensive.Orm/Orm/Upgrade/Internals/SqlWorker.cs
… Orm/Xtensive.Orm/Sql/Compiler/SqlCompiler.cs
… Orm/Xtensive.Orm/Sql/Dml/Expressions/SqlDynamicFilter.cs
… Orm/Xtensive.Orm/Sql/Info/QueryFeatures.cs
… Orm/Xtensive.Orm/Sql/SqlDml.cs
… Orm/Xtensive.Orm/Sql/SqlDriver.cs
… Orm/Xtensive.Orm/Sql/ValueTypeMapping/TypeMapper.cs
… Orm/Xtensive.Orm/Sql/ValueTypeMapping/TypeMapping.cs
… Version.props